pyproject.toml
[project]
name = "meducate-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.13.0.2",
    "openai>=2.7.2",
    "python-dotenv>=1.2.1",
]
